+++ src/lsp/cobol_preproc/preproc_engine.ml
@@ -11,6 +11,8 @@
 (*                                                                        *)
 (**************************************************************************)
 
+open EzCompat                                             (* import StringMap *)
+
 open Cobol_common.Srcloc.TYPES
 open Cobol_common.Platform.TYPES
 open Cobol_common.Srcloc.INFIX
@@ -20,20 +22,25 @@ module LIST = Cobol_common.Basics.LIST
 
 module OUT = Preproc_outputs
 module ENV = Preproc_env
+module LIT = Cobol_data.Literal
 
 (* --- *)
+
+(* TODO: fields marked with `x` below should probably be moved to a
+   `preprocessor_cold` type as they are also less likely to change as often as
+   `buff`, and `reader` *)
 
 type preprocessor =
   {
     buff: Text.t;
     reader: Src_reader.t;
-    ppstate: Preproc_state.t;
-    pplog: Preproc_trace.log;
-    diags: Preproc_diagnostics.t;
-    env: Preproc_env.t;
-    context: Preproc_logic.context;
-    rev_ignored: Text.t;     (* text accumulated when not emitting (reversed) *)
-    persist: preprocessor_persist;
+    ppstate: Preproc_state.t;                                            (* x *)
+    pplog: Preproc_trace.log;                                            (* x *)
+    diags: Preproc_diagnostics.t;                                        (* x *)
+    env: Preproc_env.t;                                                  (* x *)
+    context: Preproc_logic.context;                                      (* x *)
+    rev_ignored: Text.t;   (* text accumulated when not emitting (reversed) x *)
+    persist: preprocessor_persist;                                       (* x *)
   }
 
 (** the preprocessor state that does not change very often *)
@@ -43,10 +50,11 @@ and preprocessor_persist =
     overlay_manager: (module Src_overlay.MANAGER);
     replacing: Preproc_directives.replacing with_loc list list;
     copybooks: Cobol_common.Srcloc.copylocs;              (* opened copybooks *)
+    copybook_rev_comments: Text.comments StringMap.t;
+    copybook_lookup_config: Cobol_common.Copybook.TYPES.lookup_config;
     dialect: Cobol_config.dialect;
     source_format: Src_format.any option;  (* to keep auto-detecting on reset *)
     exec_preprocs: exec_preprocessor EXEC_MAP.t;
-    copybook_lookup_config: Cobol_common.Copybook.TYPES.lookup_config;
     platform: Cobol_common.Platform.TYPES.platform;     (* == reader.platform *)
     show_if_verbose: [`Txt | `Src] list;
   }
@@ -64,11 +72,51 @@ let add_warn lp w =
   { lp with diags = Preproc_diagnostics.add_warning w lp.diags }
 
 let position { reader; _ } = Src_reader.position reader
-let input_file { reader; _ } = Src_reader.input_file reader
+let input_filename { reader; _ } = Src_reader.input_filename reader
 let source_format { reader; _ } = Src_reader.source_format reader
 let rev_log { pplog; _ } = pplog
-let rev_comments { reader; _ } = Src_reader.rev_comments reader
+let rev_comments ({ reader; persist; _ } as lp) =
+  let rev_comments = Src_reader.rev_comments reader in
+  let map = StringMap.add "" rev_comments persist.copybook_rev_comments in
+  match input_filename lp with
+  | Some f -> StringMap.add f rev_comments map
+  | None -> map
 let rev_ignored { reader; _ } = Src_reader.rev_ignored reader
+
+let bind_78_constant lp ~loc const_name (lit: Cobol_ptree.literal with_loc) =
+  let[@local] define ?diags value =
+    let var = ENV.var' const_name in
+    let env, def = ENV.define_compilation_var var ~loc value lp.env in
+    let diags =
+      Option.fold diags
+        ~some:Preproc_diagnostics.literal_diagnostics
+        ~none:Preproc_diagnostics.none
+    and pplog =
+      Preproc_trace.var_def ~loc ~var:~&var ~def:(Compilation_var def) lp.pplog
+    in
+    add_diags { lp with env; pplog } diags
+  in
+  match ~&lit with
+  | Alphanum alphanum ->
+      define (ENV.alphanum_literal_value (alphanum &@<- lit))
+  | Boolean bool ->
+      let value = LIT.boolean (bool &@<- lit) in
+      define (ENV.boolean_literal_value value.result) ~diags:value.diags
+  | Integer integral ->
+      let fixed = Cobol_ptree.fixed_of_strings ~integral ~fractional:"0" in
+      let value = LIT.fixed (fixed &@<- lit) in
+      define (ENV.numeric_literal_value value.result) ~diags:value.diags
+  | Fixed fixed ->
+      let value = LIT.fixed (fixed &@<- lit) in
+      define (ENV.numeric_literal_value value.result) ~diags:value.diags
+  | NumFig Zero | Fig Zero ->
+      define (ENV.numeric_literal_value (LIT.fixed_zero &@<- lit))
+  | _ ->
+      add_error lp @@ Unexpected { loc = ~@lit;
+                                   stuff = Constant_literal_kind lit }
+
+let lookup_compilation_variable lp variable =
+  ENV.find_compilation_var variable lp.env
 
 (** [position_at ~line ~char pp] computes a lexing position that corresponds to
     the given line and character indexes (all starting at 0) in the input
@@ -102,8 +150,6 @@ let with_buff lp buff =
   if lp.buff == buff then lp else { lp with buff }
 let with_pplog lp pplog =
   if lp.pplog == pplog then lp else { lp with pplog }
-let with_diags_n_pplog lp diags pplog =
-  if lp.diags == diags && lp.pplog == pplog then lp else { lp with diags; pplog }
 let with_buff_n_pplog lp buff pplog =
   if lp.buff == buff && lp.pplog == pplog then lp else { lp with buff; pplog }
 let with_replacing lp replacing =
@@ -111,6 +157,9 @@ let with_replacing lp replacing =
 
 let show tag { persist = { show_if_verbose; platform; _ }; _ } =
   platform.verbosity > 0 && LIST.mem tag show_if_verbose
+
+let record_compilation_variable_substitution lp ~loc ~var ~def =
+  with_pplog lp @@ Preproc_trace.compvar_subst ~loc ~var ~def lp.pplog
 
 let source_format_config = function
   | Cobol_config.SF sf -> Some (Src_format.from_config sf)
@@ -139,10 +188,11 @@ let preprocessor input = function
             overlay_manager = (module Om);
             replacing = [];
             copybooks = Cobol_common.Srcloc.no_copy;
+            copybook_rev_comments = StringMap.empty;
+            copybook_lookup_config;
             dialect = Config.dialect;
             source_format;
             exec_preprocs;
-            copybook_lookup_config;
             platform;
             show_if_verbose = [`Src];
           };
@@ -238,9 +288,23 @@ and apply_preproc_directive ({ env; context; _ } as lp)
     if env != lp.env || diags != Preproc_diagnostics.none
     then { lp with env; diags = Preproc_diagnostics.union diags lp.diags }
     else lp
+  and new_env_n_log lp { result = (env, log); diags } =
+    if env != lp.env || diags != Preproc_diagnostics.none || log <> []
+    then { lp with
+           env;
+           diags = Preproc_diagnostics.union diags lp.diags;
+           pplog = Preproc_trace.append_entries log lp.pplog }
+    else lp
   and new_context lp { result = context; diags } =
     if context != lp.context || diags != Preproc_diagnostics.none
     then { lp with context; diags = Preproc_diagnostics.union diags lp.diags }
+    else lp
+  and new_context_n_log lp { result = (context, log); diags } =
+    if context != lp.context || diags != Preproc_diagnostics.none || log <> []
+    then { lp with
+           context;
+           diags = Preproc_diagnostics.union diags lp.diags;
+           pplog = Preproc_trace.append_entries log lp.pplog }
     else lp
   in
   match ppdir with
@@ -248,14 +312,14 @@ and apply_preproc_directive ({ env; context; _ } as lp)
     when not (Preproc_logic.emitting lp.context) ->
       lp                                                            (* ignore *)
   | Define def ->
-      new_env lp @@ Preproc_logic.on_define ~loc def ~env
+      new_env_n_log lp @@ Preproc_logic.on_define ~loc def ~env
         ~platform:lp.persist.platform
   | Define_off var ->
       new_env lp @@ Preproc_logic.on_define_off ~loc var ~env
   | If condition ->
-      new_context lp @@ Preproc_logic.on_if ~loc ~condition ~env context
+      new_context_n_log lp @@ Preproc_logic.on_if ~loc ~condition ~env context
   | Elif condition ->
-      new_context lp @@ Preproc_logic.on_elif ~loc ~condition ~env context
+      new_context_n_log lp @@ Preproc_logic.on_elif ~loc ~condition ~env context
   | Else ->
       new_context lp @@ Preproc_logic.on_else ~loc context
   | End
@@ -438,35 +502,40 @@ and do_exec ?(partial = false) lp rev_prefix exec_block suffix =
 and read_lib ({ persist = { copybook_lookup_config; platform;
                             copybooks; _ }; _ } as lp)
     loc { txtname; libname } =
-  let text, diags, pplog =
+  let text, diags, pplog, copybook_rev_comments =
     match
       platform.find_lib
         ~&txtname ?libname:~&?libname
-        ?fromfile:(input_file lp) ~lookup_config:copybook_lookup_config
+        ?fromfile:(input_filename lp) ~lookup_config:copybook_lookup_config
     with
     | Ok filename when Cobol_common.Srcloc.mem_copy filename copybooks ->
         [],
         Preproc_diagnostics.add_error
           (Cyclic_copy { copyloc = loc; filename }) lp.diags,
-        Preproc_trace.cyclic_copy ~loc ~filename lp.pplog
+        Preproc_trace.cyclic_copy ~loc ~filename lp.pplog,
+        lp.persist.copybook_rev_comments
     | Ok filename ->
         if platform.verbosity>0 then
           platform.error "Reading library `%s'@." filename;
-        let text, lp =             (* note: [lp] holds all prev and new diags *)
+        let text, lp' =             (* note: [lp] holds all prev and new diags *)
           Src_input.from ~filename ~platform ~f:begin fun input ->
             full_text                                   (* likewise for pplog *)
               (preprocessor input (`Fork (lp, loc, filename)))
               ~postproc:(Cobol_common.Srcloc.copy_from ~filename ~copyloc:loc)
           end
         in
-        text, lp.diags, Preproc_trace.copy_done ~loc ~filename lp.pplog
+        text, lp'.diags, Preproc_trace.copy_done ~loc ~filename lp'.pplog,
+        StringMap.union (fun _ x _ -> Some x (* unreachable *))
+          lp.persist.copybook_rev_comments (rev_comments lp')
     | Error lnf ->
         [],
         Preproc_diagnostics.add_error
           (Copybook_lookup_error { copyloc = Some loc; lnf }) lp.diags,
-        Preproc_trace.missing_copy ~loc ~error:lnf lp.pplog
+        Preproc_trace.missing_copy ~loc ~error:lnf lp.pplog,
+        lp.persist.copybook_rev_comments
   in
-  text, with_diags_n_pplog lp diags pplog
+  text,
+  { lp with diags; pplog; persist = { lp.persist with copybook_rev_comments } }
 
 
 and full_text ?(item = "library") ?postproc lp : Text.text * preprocessor =